🐛 FIX: Make resume ATS friendly
It sucks that good qualified applicants who could make a great fit and are extremely qualified are passed over because they don't speak the language of the ATS systems. Most of these people probably don't even know that the ATS systems exist - especially those in the older demographic. I understand that recruiters are extremely overworked with far too many resumes to look at, but there should be some sort of human element to giving each person a fair chance. I wasted about a month and a half of looking before I learned that I had to speak the computer gatekeeper language.